requirements.txt
# Web框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain和LangGraph - 使用兼容版本
langchain==0.1.4
langchain-openai==0.0.5
langchain-core==0.1.23
langgraph==0.0.20
langchain-community==0.0.16

# HTTP客户端
httpx>=0.26.0
aiohttp>=3.9.0

# 数据库（可选）
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pymysql>=1.1.0

# 工具和实用程序
python-dotenv>=1.0.0
python-multipart>=0.0.6
pyyaml>=6.0.1
nest-asyncio>=1.5.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.23.0

